Created car agent template for students
