+++ Assets/Scripts/car_agents/car_agent_template.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 527a3d4372eebd3429040010680163af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
